add validation

+++ src/main/java/task/spring/reactive/data/repository/EmployeeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import task.spring.reactive.data.model.Employee;
 
 @Repository
-public interface EmployeeRepository extends ReactiveCrudRepository<Employee, Integer> {
+public interface EmployeeRepository extends ReactiveCrudRepository<Employee, Long> {
 }
